feat: rethink of logging
main thoughts:
- there are loggers (calls of .publish() from my code)
- there are consumers (dashboard, cli, fs, etc)

key tasks of consumers:
- format message in appropriate way
- put this message to the endpoint

In case of dashboard - endpoint is runtime storage
In case of cli - endpoint is console
In case of fs - endpoint is file

So there is no "Logger" as a thing at all,
only logs publishers and logs consumers.
Consumers format logs and put to some particular place.
For convenience consumers will be named as loggers.